add test cases for PartyDao

+++ mora-insidesoccer/src/test/resources/sql/dao/is-clean.sql
@@ -1,3 +1,4 @@
 delete from login;
 
-delete from party;
+delete from party where party_owner_fk is not null;
+delete from party where party_owner_fk is null;
